refactor: extract renderer (renderCurrentPage, zoom) into js/renderer.js
