+++ js/app.js
@@ -8,9 +8,8 @@
  */
 
 import {
-  initPdfJs, loadDocument, renderPage, renderTextLayer,
-  renderThumbnail, getNextZoom, calculateFitWidth, calculateFitPage,
-  cleanupPage, getCleanupDistance,
+  initPdfJs, loadDocument,
+  renderThumbnail, getNextZoom,
 } from './pdf-engine.js';
 
 import {
@@ -26,7 +25,7 @@ import {
 
 import {
   initAnnotations, setTool, savePageAnnotations,
-  loadPageAnnotations, resizeOverlay, deleteSelected,
+  loadPageAnnotations, deleteSelected,
   updateToolOptions, getCanvas, hasAnnotations, getAnnotations, insertImage,
   undoAnnotation, redoAnnotation,
   bringToFront, sendToBack, bringForward, sendBackward,
@@ -47,14 +46,14 @@ import {
 import { icon } from './icons.js';
 
 import {
-  detectFormFields, detectFormFieldsPdfJs, renderFormOverlay, clearFormOverlay,
-  writeFormValues, hasFormFields, resetFormState,
+  detectFormFields, detectFormFieldsPdfJs, clearFormOverlay,
+  writeFormValues, resetFormState,
 } from './forms.js';
 
 import {
   buildTextIndex, clearTextIndex, searchText, findNext as findNextMatch,
   findPrevious as findPrevMatch, getMatchInfo, renderHighlights,
-  scrollToActiveHighlight, isFindOpen, setFindOpen, hasMatches,
+  scrollToActiveHighlight, isFindOpen, setFindOpen,
   augmentTextIndex, getCurrentMatchInfo, getAllMatchInfos,
   removeCurrentMatch, clearMatches,
 } from './find.js';
@@ -63,8 +62,8 @@ import { applyBatesNumbers, previewBatesLabel } from './bates.js';
 import { applyHeadersFooters, previewHeaderText } from './headers.js';
 import { openSignatureModal, closeSignatureModal, initSignatureEvents } from './signatures.js';
 import {
-  runOCR, hasOCRResults, renderOCRTextLayer, getOCRTextEntries,
-  clearOCRResults, terminateOCR, isPageScanned,
+  runOCR, hasOCRResults, getOCRTextEntries,
+  clearOCRResults, isPageScanned,
   enableCorrectionMode, disableCorrectionMode, exportOCRText, getOCRStats,
 } from './ocr.js';
 
@@ -83,10 +82,10 @@ import {
 } from './comment-summary.js';
 import { compareDocuments, generateCompareReport, renderComparisonView } from './doc-compare.js';
 import { pushDocState, undoDoc, redoDoc, canUndoDoc, canRedoDoc, clearDocHistory } from './doc-history.js';
-import { followLink, normalizeURL, extractLinksFromPage, createLinkRect } from './links.js';
+import { followLink, normalizeURL } from './links.js';
 import { getAuthorName, setAuthorName, addReply, setThreadStatus, getAllThreads, exportThreadsXFDF } from './comments.js';
 import { restoreFonts } from './font-manager.js';
-import { canUndo, canRedo, initPageState } from './history.js';
+import { canUndo, canRedo } from './history.js';
 import { enterTextEditMode, exitTextEditMode, commitTextEdits, isTextEditActive, hasTextEditChanges, enterImageEditMode, exitImageEditMode, commitImageEdits, isImageEditActive, hasImageEditChanges, canUndoImage, undoImageAction, extractImagePositions } from './text-edit.js';
 import { addExhibitStamp, setExhibitOptions, resetExhibitCount, countExistingExhibits, EXHIBIT_FORMATS } from './exhibit-stamps.js';
 import { setLabelRange, getPageLabel, getLabelRanges, clearLabels, removeLabelRange, previewLabels, LABEL_FORMATS } from './page-labels.js';
@@ -96,17 +95,13 @@ import { initMenuActions } from './menu-actions.js';
 import { parseIntegrationParams, postToCallback } from './integration.js';
 import State from './state.js';
 import { DOM, $, resolveDOMRefs } from './dom-refs.js';
+import {
+  renderCurrentPage, setZoom, setZoomThrottled, zoomIn, zoomOut,
+  fitWidth, fitPage, updateZoomDisplay, replaceIcons, setRendererCallbacks,
+  _captureScrollRatio, setPendingScrollRestore,
+} from './renderer.js';
 
 /* ═══════════════════ Initialization ═══════════════════ */
-
-/** Replace all [data-icon] elements with inline SVGs from the icon system */
-function replaceIcons() {
-  document.querySelectorAll('[data-icon]').forEach(el => {
-    const name = el.dataset.icon;
-    const size = parseInt(el.dataset.iconSize) || 16;
-    el.innerHTML = icon(name, size);
-  });
-}
 
 async function boot() {
   // Initialize centralized error handling first (before anything else)
@@ -150,6 +145,9 @@ async function boot() {
     // Listen for actual size event from menu-actions
     document.addEventListener('mudbrick:actualsize', () => setZoom(1.0));
     initModalFocusTrapping();
+
+    // Register renderer callbacks (avoids circular imports)
+    setRendererCallbacks({ updateUndoRedoButtons, refreshNotesSidebar });
 
     // Replace emoji placeholders with SVG icons
     replaceIcons();
@@ -581,176 +579,9 @@ async function openPDF(bytes, fileName, fileSize) {
 
 /* ═══════════════════ Page Rendering ═══════════════════ */
 
-// Render-pipeline state: generation counter for cancellation, debounce for navigation
-let _renderGeneration = 0;       // incremented each time we start a new render
+// Navigation debounce state
 let _navDebounceTimer = null;    // timer for debounced page navigation
 const NAV_DEBOUNCE_MS = 16;      // ~60fps batching for arrow-key holding
-
-async function renderCurrentPage() {
-  if (!State.pdfDoc) return;
-
-  // Exit edit modes when navigating to a different page (warn if unsaved)
-  if (isTextEditActive()) {
-    if (hasTextEditChanges() && !confirm('You have unsaved text edits on this page. Discard changes?')) return;
-    exitTextEditMode();
-    DOM.btnEditText.classList.remove('active');
-  }
-  if (isImageEditActive()) {
-    if (hasImageEditChanges() && !confirm('You have unsaved image edits on this page. Discard changes?')) return;
-    exitImageEditMode();
-    DOM.btnEditImage.classList.remove('active');
-  }
-
-  // Save annotations from the page we're leaving
-  if (State._prevPage && State._prevPage !== State.currentPage) {
-    savePageAnnotations(State._prevPage);
-  }
-  State._prevPage = State.currentPage;
-
-  // Bump generation counter so any in-flight render knows it has been superseded
-  const generation = ++_renderGeneration;
-
-  const result = await renderPage(
-    State.pdfDoc,
-    State.currentPage,
-    State.zoom,
-    DOM.pdfCanvas
-  );
-
-  if (!result) return; // render was cancelled by pdf-engine (a newer render started)
-
-  // If another renderCurrentPage() call started while we were awaiting, bail out
-  if (generation !== _renderGeneration) return;
-
-  const { viewport, page } = result;
-  const renderedPage = State.currentPage; // snapshot page number in case of async race
-
-  // Render text layer
-  await renderTextLayer(page, viewport, DOM.textLayer);
-  if (generation !== _renderGeneration) return;
-
-  // Store viewport and render find highlights
-  State._viewport = viewport;
-  if (hasMatches()) {
-    renderHighlights(renderedPage, DOM.textLayer, viewport);
-  }
-
-  // Inject OCR text layer if this page has been OCR'd
-  if (hasOCRResults(renderedPage)) {
-    renderOCRTextLayer(renderedPage, DOM.textLayer, viewport);
-  }
-
-  // Size the Fabric wrapper to match
-  const w = Math.floor(viewport.width);
-  const h = Math.floor(viewport.height);
-  DOM.fabricWrapper.style.width = w + 'px';
-  DOM.fabricWrapper.style.height = h + 'px';
-
-  // Resize and reload annotation overlay
-  resizeOverlay(w, h, State.zoom);
-  loadPageAnnotations(renderedPage);
-
-  // Extract existing PDF link annotations if no Fabric annotations are saved for this page
-  const annotations = getAnnotations();
-  if (!annotations[renderedPage] && State.pdfLibDoc) {
-    try {
-      const pdfPage = State.pdfLibDoc.getPage(renderedPage - 1);
-      const { height: pageH } = pdfPage.getSize();
-      const pdfLinks = extractLinksFromPage(pdfPage, pageH);
-      if (pdfLinks.length > 0) {
-        const fabCanvas = getCanvas();
-        const { width: pageW } = pdfPage.getSize();
-        const sx = w / pageW;
-        const sy = h / pageH;
-        for (const link of pdfLinks) {
-          createLinkRect(fabCanvas, link.x * sx, link.y * sy, link.width * sx, link.height * sy, {
-            linkType: link.type,
-            linkURL: link.url,
-            linkPage: link.page,
-          });
-        }
-        fabCanvas.discardActiveObject();
-        fabCanvas.renderAll();
-        savePageAnnotations(renderedPage);
-      }
-    } catch (_e) { /* ignore extraction errors */ }
-  }
-
-  // Ensure undo history has a baseline state for this page
-  const canvas = getCanvas();
-  if (canvas) {
-    initPageState(renderedPage, canvas.toJSON());
-  }
-
-  // Render form field overlays
-  clearFormOverlay();
-  if (State.formFields.length > 0 && State.pdfLibDoc) {
-    renderFormOverlay(
-      DOM.pageContainer,
-      State.formFields,
-      State.pdfLibDoc,
-      renderedPage - 1, // 0-based page index
-      State.zoom,
-      { width: viewport.width / State.zoom, height: viewport.height / State.zoom }
-    );
-  }
-
-  // Size the page container
-  DOM.pageContainer.style.width = Math.floor(viewport.width) + 'px';
-  DOM.pageContainer.style.height = Math.floor(viewport.height) + 'px';
-
-  // Restore scroll position after zoom-triggered resize
-  if (_pendingScrollRestore) {
-    const sr = _pendingScrollRestore;
-    _pendingScrollRestore = null;
-    const el = DOM.canvasArea;
-    if (sr.type === 'point') {
-      // Cursor-anchored: keep the same page point under the cursor
-      el.scrollLeft = sr.pageX * State.zoom - sr.clientX;
-      el.scrollTop = sr.pageY * State.zoom - sr.clientY;
-    } else {
-      // Ratio-based: preserve relative scroll position
-      const maxX = el.scrollWidth - el.clientWidth;
-      const maxY = el.scrollHeight - el.clientHeight;
-      el.scrollLeft = sr.rx * maxX;
-      el.scrollTop = sr.ry * maxY;
-    }
-  }
-
-  // Refresh Notes sidebar
-  refreshNotesSidebar();
-
-  // Cleanup distant pages using a distance that adapts to memory pressure
-  const _cleanupDist = getCleanupDistance();
-  for (let i = 1; i <= State.totalPages; i++) {
-    if (Math.abs(i - renderedPage) > _cleanupDist) {
-      cleanupPage(State.pdfDoc, i);
-    }
-  }
-
-  updateUndoRedoButtons();
-
-  // Pre-warm adjacent pages during idle time so next/prev navigation feels instant.
-  // Uses a detached off-screen canvas so the main canvas is never disturbed.
-  if (typeof requestIdleCallback === 'function') {
-    requestIdleCallback(() => {
-      if (!State.pdfDoc || generation !== _renderGeneration) return;
-      const preWarm = (pNum) => {
-        if (pNum < 1 || pNum > State.totalPages) return;
-        const offCanvas = document.createElement('canvas');
-        State.pdfDoc.getPage(pNum).then(pg => {
-          const vp = pg.getViewport({ scale: State.zoom });
-          offCanvas.width = Math.floor(vp.width);
-          offCanvas.height = Math.floor(vp.height);
-          const ctx = offCanvas.getContext('2d');
-          return pg.render({ canvasContext: ctx, viewport: vp }).promise;
-        }).catch(() => {}); // best-effort warm-up; ignore errors
-      };
-      preWarm(renderedPage - 1);
-      preWarm(renderedPage + 1);
-    }, { timeout: 2000 });
-  }
-}
 
 /* ═══════════════════ Navigation ═══════════════════ */
 
@@ -798,94 +629,6 @@ function updatePageNav() {
   DOM.btnPrev.disabled = atFirst;
   DOM.btnNext.disabled = atLast;
   if (DOM.btnLast) DOM.btnLast.disabled = atLast;
-}
-
-/* ═══════════════════ Zoom ═══════════════════ */
-
-// Zoom throttle state — ensures Ctrl+scroll updates happen at most once per frame
-let _zoomRafPending = false;
-let _pendingZoom = null;
-const ZOOM_THROTTLE_MS = 16; // ~60fps
-
-// Scroll restore state — preserves viewport position across zoom changes.
-// Two modes:
-//   { type: 'ratio', rx, ry }                          — ratio-based (button zoom)
-//   { type: 'point', pageX, pageY, clientX, clientY }  — cursor-anchored (Ctrl+scroll)
-let _pendingScrollRestore = null;
-
-function _captureScrollRatio() {
-  const el = DOM.canvasArea;
-  const maxX = el.scrollWidth - el.clientWidth;
-  const maxY = el.scrollHeight - el.clientHeight;
-  _pendingScrollRestore = {
-    type: 'ratio',
-    rx: maxX > 0 ? el.scrollLeft / maxX : 0.5,
-    ry: maxY > 0 ? el.scrollTop / maxY : 0.5,
-  };
-}
-
-function setZoom(newZoom) {
-  _captureScrollRatio();
-  State.zoom = Math.max(0.25, Math.min(5.0, newZoom));
-  updateZoomDisplay();
-  renderCurrentPage();
-}
-
-/**
- * Throttled zoom setter for Ctrl+scroll — batches rapid wheel events to 60fps.
- * The display is updated immediately for smooth visual feedback; the expensive
- * re-render only fires once per animation frame.
- */
-function setZoomThrottled(newZoom) {
-  // Scroll restore is set by the wheel handler (cursor-anchored), not here
-  _pendingZoom = Math.max(0.25, Math.min(5.0, newZoom));
-  State.zoom = _pendingZoom;   // update state immediately so display feels snappy
-  updateZoomDisplay();
-
-  if (!_zoomRafPending) {
-    _zoomRafPending = true;
-    requestAnimationFrame(() => {
-      _zoomRafPending = false;
-      if (_pendingZoom !== null) {
-        const zoom = _pendingZoom;
-        _pendingZoom = null;
-        State.zoom = zoom;
-        renderCurrentPage();
-      }
-    });
-  }
-}
-
-function zoomIn() { setZoom(getNextZoom(State.zoom, 1)); }
-function zoomOut() { setZoom(getNextZoom(State.zoom, -1)); }
-
-function fitWidth() {
-  if (!State.pdfDoc) return;
-  State.pdfDoc.getPage(State.currentPage).then(page => {
-    const viewport = page.getViewport({ scale: 1 });
-    const container = DOM.canvasArea;
-    const newZoom = calculateFitWidth(viewport.width, container.clientWidth);
-    setZoom(newZoom);
-  }).catch(() => {});
-}
-
-function fitPage() {
-  if (!State.pdfDoc) return;
-  State.pdfDoc.getPage(State.currentPage).then(page => {
-    const viewport = page.getViewport({ scale: 1 });
-    const container = DOM.canvasArea;
-    const newZoom = calculateFitPage(
-      viewport.width, viewport.height,
-      container.clientWidth, container.clientHeight
-    );
-    setZoom(newZoom);
-  }).catch(() => {});
-}
-
-function updateZoomDisplay() {
-  const pct = Math.round(State.zoom * 100) + '%';
-  DOM.zoomBtn.textContent = pct;
-  DOM.statusZoom.textContent = pct;
 }
 
 /* ═══════════════════ Thumbnails ═══════════════════ */
@@ -4044,7 +3787,7 @@ function wireEvents() {
       const pageX = (DOM.canvasArea.scrollLeft + clientX) / oldZoom;
       const pageY = (DOM.canvasArea.scrollTop + clientY) / oldZoom;
 
-      _pendingScrollRestore = { type: 'point', pageX, pageY, clientX, clientY };
+      setPendingScrollRestore({ type: 'point', pageX, pageY, clientX, clientY });
       setZoomThrottled(next);
     }
   }, { passive: false });
